feat: apply access control to all earn & portfolio
