+++ src/app/ui/earn/EarnsPage.tsx
@@ -1,8 +1,5 @@
-'use client';
-
 import type { FC } from 'react';
 
-import { Gatekeeper } from '../gatekeeper/Gatekeeper';
 import { EarnOpportunitiesAll } from './EarnOpportunitiesAll/EarnOpportunitiesAll';
 import { EarnTopOpportunities } from './EarnTopOpportunities';
 
@@ -10,9 +7,9 @@ interface EarnsPageProps {}
 
 export const EarnsPage: FC<EarnsPageProps> = () => {
   return (
-    <Gatekeeper flag="hasEarn">
+    <>
       <EarnTopOpportunities />
       <EarnOpportunitiesAll />
-    </Gatekeeper>
+    </>
   );
 };
